WIP for Meteor 1.3.
