+++ .gitignore
@@ -1,2 +1,3 @@
 *.cached
 *.map
+node_modules/
