--- lib/react-router-ssr.js
@@ -1,1 +1,0 @@
-ReactRouterSSR = {};
