+++ package.json
@@ -22,5 +22,15 @@
   "bugs": {
     "url": "https://github.com/thereactivestack/meteor-react-router-ssr/issues"
   },
-  "homepage": "https://github.com/thereactivestack/meteor-react-router-ssr#readme"
+  "homepage": "https://github.com/thereactivestack/meteor-react-router-ssr#readme",
+  "dependencies": {
+    "cheerio": "^0.20.0",
+    "cookie-parser": "^1.4.1",
+    "promise": "^7.1.1",
+    "react": "^0.14.8",
+    "react-dom": "^0.14.8",
+    "react-helmet": "^3.0.0",
+    "react-router": "^2.0.1",
+    "underscore": "^1.8.3"
+  }
 }
